backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / LangChain
langchain==0.3.0
langchain-anthropic==0.3.0
langchain-community==0.3.0
faiss-cpu==1.8.0

# AWS
boto3==1.35.0

# Redis
redis==5.1.0

# PDF Processing
pypdf==4.3.0

# Utilities
pydantic-settings==2.5.0
python-dotenv==1.0.1

# Testing
pytest==8.3.0
pytest-cov==5.0.0
pytest-asyncio==0.24.0
httpx==0.27.0
moto[s3]==5.0.0
fakeredis==2.25.0
